+++ mypyc/irbuild/expression.py
@@ -21,7 +21,7 @@ from mypyc.ops import (
 from mypyc.ops_primitive import name_ref_ops
 from mypyc.primitives.misc_ops import new_slice_op, iter_op, ellipsis_op, type_op
 from mypyc.ops_list import new_list_op, list_append_op, list_extend_op
-from mypyc.ops_tuple import list_tuple_op
+from mypyc.primitives.tuple_ops import list_tuple_op
 from mypyc.primitives.dict_ops import new_dict_op, dict_set_item_op
 from mypyc.primitives.set_ops import new_set_op, set_add_op, set_update_op
 from mypyc.irbuild.specialize import specializers
